requirements.txt
django
djangorestframework
django-filter
requests
pytelegrambotapi
beautifulsoup4
lxml